flytt utils-fil til samme mappe som kode og rename
